finished pagination, and worked on activities view

+++ resources/views/activities/index.blade.php
@@ -2,7 +2,19 @@
 
 @section('content')
     <div class="container">
-        <a href="{{ route('activities.create') }}" class="btn btn-primary">Creer une nouvelle Activité</a>
+        <div style="text-align: center">
+            <ul class="nav nav-pills" style="display: inline-block">
+                @if($nav =='coming')
+                    <li class="active"><a href="{{route('activities.index', ['nav' => 'coming'])}}">Evenements à venir</a></li>
+                    <li><a href="{{route('activities.index', ['nav' => 'past'])}}">Evenements passés</a></li>
+                @else
+                    <li><a href="{{route('activities.index', ['nav' => 'coming'])}}">Evenements à venir</a></li>
+                    <li class="active"><a href="{{route('activities.index', ['nav' => 'past'])}}">Evenements passés</a></li>
+                @endif
+
+            </ul>
+        </div>
+
         @foreach($activities as $activity)
             <div class="well" style="margin-top: 20px; padding: 5px">
                 <div class="row">
@@ -22,22 +34,26 @@
                         @if(sizeof($dates[$activity->id])==1)
                             <h4>date de l'évènement :</h4>
                             <h4>{{ $dates[$activity->id][0]->date }}</h4>
-                            @if(isset($likedates[$activity->id]))
-                                {{ Form::open(['method' => 'DELETE', 'route' => ['likeDates.destroy', $likedates[$activity->id]]]) }}
-                                {{ Form::submit('je ne participe plus', ['class' => 'btn btn-default']) }}
-                                {{ Form::close() }}
-                            @else
-                                {{ Form::open(['route' => ['likeDates.store']]) }}
-                                {{ Form::hidden('date_id', $dates[$activity->id][0]->id) }}
-                                {{ Form::submit('je participe !', ['class' => 'btn btn-default']) }}
-                                {{ Form::close() }}
+                            @if($nav == 'coming')
+                                @if(isset($likedates[$activity->id]))
+                                    {{ Form::open(['method' => 'DELETE', 'route' => ['likeDates.destroy', $likedates[$activity->id]]]) }}
+                                    {{ Form::submit('je ne participe plus', ['class' => 'btn btn-danger']) }}
+                                    {{ Form::close() }}
+                                @else
+                                    {{ Form::open(['route' => ['likeDates.store']]) }}
+                                    {{ Form::hidden('date_id', $dates[$activity->id][0]->id) }}
+                                    {{ Form::submit('je participe !', ['class' => 'btn btn-success']) }}
+                                    {{ Form::close() }}
+                                @endif
                             @endif
                         @else
                             <table class="table table-hover">
                                 <thead>
                                 <tr>
                                     <th style="text-align: center">dates proposées</th>
-                                    <th style="text-align: center">vote</th>
+                                    @if($nav == 'coming')
+                                        <th style="text-align: center">vote</th>
+                                    @endif
                                 </tr>
                                 </thead>
                                 <tbody>
@@ -49,20 +65,22 @@
                                 @for($i=0; $i< sizeof($dates[$activity->id]); $i++)
                                     <tr>
                                         <td>{{ $dates[$activity->id][$i]->date }}</td>
-                                        <td>
-                                            @if($liked == $dates[$activity->id][$i]->id)
-                                                <span class="glyphicon glyphicon-ok"></span>
-                                            @else
-                                                @if($liked != -1)
-                                                    {{ Form::open(['method' => 'PUT', 'route' => ['likeDates.update', $likedates[$activity->id]]]) }}
+                                        @if($nav == 'coming')
+                                            <td>
+                                                @if($liked == $dates[$activity->id][$i]->id)
+                                                    <span class="glyphicon glyphicon-ok"></span>
                                                 @else
-                                                    {{ Form::open(['route' => 'likeDates.store']) }}
+                                                    @if($liked != -1)
+                                                        {{ Form::open(['method' => 'PUT', 'route' => ['likeDates.update', $likedates[$activity->id]]]) }}
+                                                    @else
+                                                        {{ Form::open(['route' => 'likeDates.store']) }}
+                                                    @endif
+                                                    {{ Form::hidden('date_id' , $dates[$activity->id][$i]->id) }}
+                                                    {{Form::button('<i class="glyphicon glyphicon-plus"></i>', array('type' => 'submit', 'class' => 'btn btn-success btn-xs'))}}
+                                                    {{ Form::close() }}
                                                 @endif
-                                                {{ Form::hidden('date_id' , $dates[$activity->id][$i]->id) }}
-                                                {{Form::button('<i class="glyphicon glyphicon-plus"></i>', array('type' => 'submit', 'class' => 'btn btn-default btn-xs'))}}
-                                                {{ Form::close() }}
-                                            @endif
-                                        </td>
+                                            </td>
+                                        @endif
                                     </tr>
                                 @endfor
                                 </tbody>
@@ -72,6 +90,6 @@
                 </div>
             </div>
         @endforeach
-        <div class="pagination"> {{ $activities->links() }}</div>
+        <div class="pagination"> {{ $Adates->links() }}</div>
     </div>
 @endsection
